pyproject.toml
[tool.poetry]
name = "cms-nbi-client"
version = "2.0.0"
description = "Modern async Python client for Calix Management System NBI"
authors = ["somenetworking <andrewshea06@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"
python = "^3.9"
packages = [{include = "cmsnbiclient", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.9.0"
pydantic = "^2.5.0"
structlog = "^23.2.0"
xmltodict = "^0.13.0"
defusedxml = "^0.7.1"
cryptography = "^41.0.0"
python-keyring = "^24.3.0"
tenacity = "^8.2.0"
aiocache = "^0.12.0"
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.21.0"
opentelemetry-sdk = "^1.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^23.12.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.7.0"
pre-commit = "^3.6.0"
mkdocs = "^1.5.0"
mkdocs-material = "^9.5.0"
locust = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --cov=cmsnbiclient --cov-report=html"
testpaths = ["tests"]
asyncio_mode = "auto"